rename the filename html-parser to sfc-parser
