+++ src/compiler/parser/sfc-parser.js
